+++ crates/nargo_cli/tests/prove_and_verify.rs
@@ -1,14 +1,18 @@
+use tempdir::TempDir;
+
 use std::collections::BTreeMap;
 use std::fs;
 
 mod tests {
+    use std::path::Path;
     use std::path::PathBuf;
 
     use super::*;
 
-    fn load_conf(conf_path: &str) -> BTreeMap<String, Vec<String>> {
-        // Parse config.toml into a BTreeMap, do not fail if config file does not exist.
-        let mut conf_data = match toml::from_str(conf_path) {
+    fn load_conf(conf_path: &Path) -> BTreeMap<String, Vec<String>> {
+        let config_str = std::fs::read_to_string(conf_path).unwrap();
+
+        let mut conf_data = match toml::from_str(&config_str) {
             Ok(t) => t,
             Err(_) => BTreeMap::from([
                 ("exclude".to_string(), Vec::new()),
@@ -24,6 +28,24 @@ mod tests {
         conf_data
     }
 
+    /// Copy files from source to destination recursively.
+    pub fn copy_recursively(
+        source: impl AsRef<Path>,
+        destination: impl AsRef<Path>,
+    ) -> std::io::Result<()> {
+        fs::create_dir_all(&destination)?;
+        for entry in fs::read_dir(source)? {
+            let entry = entry?;
+            let filetype = entry.file_type()?;
+            if filetype.is_dir() {
+                copy_recursively(entry.path(), destination.as_ref().join(entry.file_name()))?;
+            } else {
+                fs::copy(entry.path(), destination.as_ref().join(entry.file_name()))?;
+            }
+        }
+        Ok(())
+    }
+
     #[test]
     fn noir_integration() {
         // Try to find the directory that Cargo sets when it is running; otherwise fallback to assuming the CWD
@@ -36,33 +58,49 @@ mod tests {
         let config_path = test_data_dir.join("config.toml");
 
         //load config.tml file from test_data directory
-        let config_path = std::fs::read_to_string(config_path).unwrap();
         let config_data: BTreeMap<String, Vec<String>> = load_conf(&config_path);
 
-        for c in fs::read_dir(test_data_dir).unwrap().flatten() {
-            if let Ok(test_name) = c.file_name().into_string() {
-                if c.path().is_dir() && !config_data["exclude"].contains(&test_name) {
-                    println!("Running test {test_name:?}");
-                    let verified = std::panic::catch_unwind(|| {
-                        nargo_cli::cli::prove_and_verify("pp", &c.path(), false)
-                    });
+        // Copy all the test cases into a temp dir so we don't leave artifacts around.
+        let tmp_dir = TempDir::new("p_and_v_tests").unwrap();
+        copy_recursively(test_data_dir, &tmp_dir)
+            .expect("failed to copy test cases to temp directory");
 
-                    let r = match verified {
-                        Ok(result) => result,
-                        Err(_) => {
-                            panic!("\n\n\nPanic occurred while running test {:?} (ignore the following panic)", c.file_name());
-                        }
-                    };
+        let test_case_dirs =
+            fs::read_dir(&tmp_dir).unwrap().flatten().filter(|c| c.path().is_dir());
 
-                    if config_data["fail"].contains(&test_name) {
-                        assert!(!r, "{:?} should not succeed", c.file_name());
-                    } else {
-                        assert!(r, "verification fail for {:?}", c.file_name());
-                    }
-                } else {
-                    println!("Ignoring test {test_name:?}");
+        for test_dir in test_case_dirs {
+            let test_name =
+                test_dir.file_name().into_string().expect("Directory can't be converted to string");
+            let test_program_dir = &test_dir.path();
+
+            if config_data["exclude"].contains(&test_name) {
+                println!("Skipping test {test_name}");
+                continue;
+            }
+
+            println!("Running test {test_name}");
+
+            let verified = std::panic::catch_unwind(|| {
+                nargo_cli::cli::prove_and_verify("pp", test_program_dir, false)
+            });
+
+            let r = match verified {
+                Ok(result) => result,
+                Err(_) => {
+                    panic!(
+                        "\n\n\nPanic occurred while running test {test_name} (ignore the following panic)"
+                    );
                 }
+            };
+
+            if config_data["fail"].contains(&test_name) {
+                assert!(!r, "{:?} should not succeed", test_name);
+            } else {
+                assert!(r, "verification fail for {:?}", test_name);
             }
         }
+
+        // Ensure that temp dir remains alive until all tests have run.
+        drop(tmp_dir);
     }
 }