+++ crates/noirc_evaluator/src/ssa_refactor/ir/types.rs
@@ -1,0 +1,23 @@
+/// A numeric type in the Intermediate representation
+/// Note: we class NativeField as a numeric type
+/// though we also apply limitations to it, such as not
+/// being able to compare two native fields, whereas this is
+/// something that you can do with a signed/unsigned integer.
+///
+/// Fields do not have a notion of ordering, so this distinction
+/// is reasonable.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+pub(crate) enum NumericType {
+    Signed { bit_size: u32 },
+    Unsigned { bit_size: u32 },
+    NativeField,
+}
+
+/// All types representable in the IR.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+pub(crate) enum Type {
+    /// Represents numeric types in the IR, including field elements
+    Numeric(NumericType),
+    /// The Unit type with a single value
+    Unit,
+}