+++ crates/noirc_evaluator/src/ssa_refactor/ir/value.rs
@@ -1,0 +1,19 @@
+use super::{instruction::InstructionId, map::Id, types::Type};
+
+pub(crate) type ValueId = Id<Value>;
+
+/// Value is the most basic type allowed in the IR.
+/// Transition Note: A Id<Value> is similar to `NodeId` in our previous IR.
+#[derive(Debug, PartialEq, Eq, Hash, Clone)]
+pub(crate) enum Value {
+    /// This value was created due to an instruction
+    ///
+    /// instruction -- This is the instruction which defined it
+    /// typ -- This is the `Type` of the instruction
+    /// position -- Returns the position in the results
+    /// vector that this `Value` is located.
+    /// Example, if you add two numbers together, then the resulting
+    /// value would have position `0`, the typ would be the type
+    /// of the operands, and the instruction would map to an add instruction.
+    Instruction { typ: Type, position: u16, instruction: InstructionId },
+}